third_party/WebKit/LayoutTests/http/tests/security/xss-DENIED-xsl-document-securityOrigin-expected.txt

CONSOLE MESSAGE: line 32: PASS: Caught exception while trying to access victim's properties.
This test passes if it doesn't alert the contents of innocent-victim.html.  
